+++ PharmacySystem/CompositionRoot.cs
@@ -90,6 +90,9 @@ namespace PharmacySystem
         public static SalePresenter CreateSalePresenter(ISaleView view, int idPerson) =>
             new SalePresenter(view, _saleService, _productService, _storeService, idPerson);
 
+        public static CreditNotePresenter CreateCreditNotePresenter(ICreditNoteView view) =>
+            new CreditNotePresenter(view, _saleService, MainForm.Session, MainForm.oPerson?.idPerson ?? 0);
+
         public static ReportPresenter CreateReportPresenter(IReportView view) =>
             new ReportPresenter(view, _supplierService, _categoryService, _saleService, _purchaseService, _productService, _notificationConfigService, MainForm.Session);